requirements.txt
flask
google-cloud-spanner
google-cloud-bigquery
google-cloud-pubsub
google-cloud-aiplatform
pandas
scikit-learn
nltk